Refactor zero_suppress_waveform_truth()
Rewrote the function to reduce memory allocations and change the method
for getting indices. Using some process of elimination, the
np.ndenumerate() call coupled with the list comprehension was causing a
huge slowdown.

Total run time for larnd-sim to process 25 events was reduced from ~175s
to ~100s with these changes.