Add report sub-command to lar_runner; adjust nvtx ranges
Added a report option to lar_runner to run `nsys stats` on profile
output. If running the `nvtx_sum` report, it creates an additional csv
file with a new time column to calculate the fraction of the total
runtime for each nvtx range.

Also added optional color arguments to several nvtx ranges. These colors
show up when viewing the profile in nsys-ui.