Fix issue when performing regular (non-editable) install
Found an issue where the 'consts' folder was not being copied /
installed when running 'pip install .' as a non-editable installation.
This was probably due to an error translating the original setup.py to
the pyproject.toml format.

Changed the pyproject toml to auto-find all the packages within
larndsim.

Also edited how the license was specified to fix a deprecation warning.