Update Perlmutter workflow to run on PR merge; updates to lar_runner
Updated the Perlmutter GitHub workflow to run when a PR is closed by
merging to develop (and only develop).

Also changed the output filename within the workflow to be based on the
branch, commit SHA, and time to create unique filenames and track when
and what they were run from.

Updated lar_runner.py to better craft output file names. It now checks
if the 'hdf5' extension is present for larndsim output files. When
specifying an output name with nsys or ncu it also changes the larnd-sim
file name to match.