Improve lar_runner file handling; updated runner.yaml defaults
Several improvements to how lar_runner handles files for running
profiles and creating reports.

Both `nsys` and `ncu` will now create the output directory if it does not
exis, and now properly delete the existing larnd-sim output if --force
was specified.

If the profile was initially not found, `report` will prepend the nsys
output directory to the name and use that.

Updated the runner.yaml `nsys` and `ncu` output directories and
increased larnd-sim events to 75 (from 50).