Add reduced precision light LUT for 2x2/FSD/ND-LAr
Adds new configuration options for using a reduced precision (FP16)
light lookup table to save GPU memory when running. The reduced
precision is only applied to 'time_dist' and the 'vis' and 't0' arrays
remain at FP32.

Added an explicit cast in light_sim.py to account for using the FP16
version of the LUT.